p2p/{peer,protocol}: make `snappy_support` a property.
When upgrading from no-compression to with-compression, the whole
[P2P]Protocol was re-initialised, churning objects on every connection,
even if the remote only wants to say "too many peers, bye".

Also, `snappy_support` is passed around as an argument (a lot!) -
and it should not be: it's a property estabisished by the lowest-level
interchange (handshake, Hello) of the DEVp2p protocol, and should
be "inherited" from then until the end of session.

This commit leaves it (along with command ID offset) as a parameter
to all instances of `Command`s, because that's a somewhat different
issue. These are only ballast in the code base: what their values
should be can always be determined by the protocol that "bundles"
the commands.

TODO: Address the issue described in above paragraph.