p2p/peer: remove snappy_support from select_sub_protocol().
Leaving it there suggests that it is a choice, but it is not.

As described in previous commit, there is also no way to "downgrade" a
sub-protocol to "no compression" if the base-protocol DEVp2p handshake
suggests both sides support DEVp2p v5: broadcasting v5 equates desire to
use compression.